+++ src/pages/private/admin/Setting/Portafolio/Servicios/Servicios.jsx
@@ -14,6 +14,7 @@ import { getInfoCategoria } from "../utilsPortafolio";
 import { useDispatch, useSelector } from "react-redux";
 import { deleteServicio } from "../../../../../../redux/actions/aServicios";
 import { Notify } from "../../../../../../utils/notify/Notify";
+import { formatThousandsSeparator } from "../../../../../../utils/functions";
 
 const Servicios = () => {
   const [infoServicios, setInfoServicios] = useState([]);
@@ -65,9 +66,7 @@ const Servicios = () => {
           placeholder: "",
         },
         Cell: ({ cell }) => (
-          <Box>
-            {simboloMoneda} {cell.getValue()}
-          </Box>
+          <Box>{formatThousandsSeparator(cell.getValue(), true)}</Box>
         ),
       },
       {
@@ -89,6 +88,8 @@ const Servicios = () => {
   };
 
   const handleDeleteService = (id) => {
+    let confirmationEnabled = true;
+
     modals.openConfirmModal({
       title: "Eliminacion de Servicio",
       centered: true,
@@ -99,9 +100,12 @@ const Servicios = () => {
       confirmProps: { color: "red" },
 
       onConfirm: () => {
-        dispatch(deleteServicio(id));
-        Notify("Eliminacion Exitosa", "", "success");
-        handleCloseAction();
+        if (confirmationEnabled) {
+          confirmationEnabled = false;
+          dispatch(deleteServicio(id));
+          Notify("Eliminacion Exitosa", "", "success");
+          handleCloseAction();
+        }
       },
     });
   };
